requirements.txt
absl-py==0.7.1
astor==0.8.0
backcall==0.1.0
cycler==0.10.0
decorator==4.4.0
gast==0.2.2
grpcio==1.22.0
h5py==2.9.0
imageio==2.5.0
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.14.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.0.3
networkx==2.3
numpy==1.14.5
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
progressbar==2.5
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
PyWavelets==1.0.3
PyYAML==5.1.2
scikit-image==0.15.0
scipy==1.3.0
six==1.12.0
tensorboard==1.10.0
tensorflow-gpu==1.10.0
termcolor==1.1.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.15.5
